0.2

+++ code_for_DESIRE/code/DESIRE_model.py
@@ -313,7 +313,7 @@ class RefineModel(nn.Module):
       # sp: tensor(K,batch_size,6*6,48)
       sp = torch.zeros((self.K,self.batch_size,self.social_pooling_size[0]*self.social_pooling_size[1],hidden.shape[1]), device=torch.device(self.device))
       # sp_c: count the numbers in (K,batch_size,6*6)
-      sp_c = torch.zeros((self.K,self.batch_size,self.social_pooling_size[0]*self.social_pooling_size[1]), device=torch.device(self.device),dtype=torch.int32).detach()
+      sp_c = torch.zeros((self.K,self.batch_size,int(self.social_pooling_size[0]*self.social_pooling_size[1])), device=torch.device(self.device),dtype=torch.int32).detach()
       sp_one = torch.ones_like(sp_c)
       for i in range(loc_others.shape[0]):
         # loc:tensor(k,batch_size,2)
